// src/Frontend/src/app/server/model/applicationModelsApiModelsApiRequirementSpecificationTypes.ts
/**
 * WebApi
 * No description provided (generated by Swagger Codegen https://github.com/swagger-api/swagger-codegen)
 *
 * OpenAPI spec version: 1.0.0
 * 
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */

export type ApplicationModelsApiModelsApiRequirementSpecificationTypes = 'Base' | 'Person' | 'Material' | 'Money';

export const ApplicationModelsApiModelsApiRequirementSpecificationTypes = {
    Base: 'Base' as ApplicationModelsApiModelsApiRequirementSpecificationTypes,
    Person: 'Person' as ApplicationModelsApiModelsApiRequirementSpecificationTypes,
    Material: 'Material' as ApplicationModelsApiModelsApiRequirementSpecificationTypes,
    Money: 'Money' as ApplicationModelsApiModelsApiRequirementSpecificationTypes
};