// src/Frontend/src/app/server/model/applicationModelsApiModelsApiContactSpecificationPhoneNumber.ts
/**
 * WebApi
 * No description provided (generated by Swagger Codegen https://github.com/swagger-api/swagger-codegen)
 *
 * OpenAPI spec version: 1.0.0
 * 
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */
import { ApplicationModelsApiModelsApiContactSpecification } from './applicationModelsApiModelsApiContactSpecification';
import { DomainDataTypesPhoneNumber } from './domainDataTypesPhoneNumber';

export interface ApplicationModelsApiModelsApiContactSpecificationPhoneNumber extends ApplicationModelsApiModelsApiContactSpecification { 
    phoneNumber?: DomainDataTypesPhoneNumber;
}